+++ prisma/seed/seed.ts
@@ -1,75 +1,10 @@
 import { PrismaClient } from "@prisma/client";
-import { assert } from "../../src/utils/asserts";
-import {
-  DELTAHACKS_APPLICATION_FORM_CONFIG,
-  FORM_STRUCTURES,
-  FormStructure,
-} from "./data";
+import { createForm } from "../../src/server/router/formBuilder";
+import { DELTAHACKS_APPLICATION_FORM_CONFIG, FORM_STRUCTURES } from "./data";
 
 const prisma = new PrismaClient({
   log: ["query", "error", "warn"],
 });
-
-const createFormStructure = async (formStructure: FormStructure) => {
-  await prisma.formStructure.createMany({
-    data: { id: formStructure.id },
-    skipDuplicates: true,
-  });
-
-  await prisma.questionCategory.createMany({
-    data: formStructure.categories.map(({ name }, i) => ({
-      formStructureId: formStructure.id,
-      name,
-      formPosition: i,
-    })),
-    skipDuplicates: true,
-  });
-
-  const categories = await prisma.questionCategory.findMany({
-    where: {
-      formStructureId: formStructure.id,
-    },
-    select: {
-      id: true,
-      name: true,
-    },
-  });
-
-  assert(
-    categories.length >= formStructure.categories.length,
-    "All categories should've been added before"
-  );
-
-  const categoryNameToIdMap = new Map(
-    categories.map(({ id, name }) => [name, id])
-  );
-
-  await Promise.all(
-    formStructure.categories.map(async ({ name: categoryName, questions }) => {
-      const categoryId = categoryNameToIdMap.get(categoryName);
-      assert(categoryId);
-
-      try {
-        await prisma.question.createMany({
-          data: questions.map((question, i) => {
-            return {
-              statement: question,
-              categoryPosition: i,
-              categoryId,
-            };
-          }),
-          skipDuplicates: true,
-        });
-      } catch (e) {
-        console.error(
-          `Error occured while adding the following: categoryName: ${categoryName}, categoryId: ${categoryId}, questions:`,
-          questions
-        );
-        throw e;
-      }
-    })
-  );
-};
 
 async function main() {
   if (process.env.NODE_ENV === "production") {
@@ -87,7 +22,7 @@ async function main() {
 
   await Promise.all(
     FORM_STRUCTURES.map(async (structure) => {
-      await createFormStructure(structure);
+      await createForm(prisma, structure);
     })
   );
 }
